Move Phase enum to Game
